# backend/docs/api.yaml
openapi: 3.0.3
info:
  title: TON Empire Game API
  description: |
    REST API для игры TON Empire - социальной стратегии на блокчейне TON.
    
    ## Аутентификация
    API использует JWT токены для аутентификации. Токены получаются через Telegram Mini App.
    
    ## Rate Limiting
    - Аутентификация: 10 запросов в минуту
    - Game API: 100 запросов в минуту
    - WebSocket: без ограничений
    
    ## Коды ошибок
    - 400: Некорректный запрос
    - 401: Не авторизован
    - 403: Доступ запрещен
    - 404: Ресурс не найден
    - 429: Превышен лимит запросов
    - 500: Внутренняя ошибка сервера
  version: 1.0.0
  contact:
    name: TON Empire Development Team
    email: dev@ton-empire.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.ton-empire.com
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /api/users/me:
    get:
      tags:
        - Users
      summary: Получить профиль текущего пользователя
      responses:
        '200':
          description: Профиль пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'

    put:
      tags:
        - Users
      summary: Обновить профиль пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  minLength: 3
                  maxLength: 32
                first_name:
                  type: string
                  maxLength: 64
                last_name:
                  type: string
                  maxLength: 64
      responses:
        '200':
          description: Профиль обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/users/{userId}:
    get:
      tags:
        - Users
      summary: Получить профиль пользователя по ID
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Профиль пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicUser'
        '404':
          $ref: '#/components/responses/NotFound'

  # District Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        telegram_id:
          type: string
        username:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        level:
          type: integer
          minimum: 1
        experience:
          type: integer
          minimum: 0
        power:
          type: integer
          minimum: 0
        battle_stats:
          type: object
          properties:
            wins:
              type: integer
            losses:
              type: integer
            rating:
              type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        last_active_at:
          type: string
          format: date-time
    PublicUser:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        level:
          type: integer
        power:
          type: integer
        battle_stats:
          type: object
          properties:
            wins:
              type: integer
            losses:
              type: integer
            rating:
              type: integer
        guild:
          type: object
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
            tag:
              type: string
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Код ошибки
        message:
          type: string
          description: Описание ошибки
        details:
          type: object
          description: Дополнительная информация об ошибке
  responses:
    BadRequest:
      description: Некорректный запрос
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "bad_request"
            message: "Invalid request parameters"
    Unauthorized:
      description: Не авторизован
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "unauthorized"
            message: "Authentication required"
...
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "not_found"
            message: "Resource not found"
...